requirements_new.txt
# Telegram Bot
python-telegram-bot==20.7

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Utilities
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0

# Development
python-dateutil==2.8.2 